moved and renamed circleVisualisation and Path
