+++ src/components/d3/CircleVisualisation.vue
@@ -91,7 +91,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import store from '../store'
+import store from '../../store'
 
 export default Vue.extend({
   name: 'CircleVisualisation',
